+++ docs/history.rst
@@ -6,6 +6,7 @@ Release History
 `Next Release`_
 ---------------
 - Make shutdown timings configurable.
+- Add :class:`sprockets.http.testing.SprocketsHttpTestCase`.
 
 `2.0.1`_ (5 Mar 2019)
 ----------------------